Got datepicker functional, added total income query, and toal income by family added confirmation buttons for delete and edit

+++ src/app/income/income.component.ts
@@ -20,7 +20,7 @@ export class IncomeComponent implements OnInit {
 
   constructor(private router: Router, private familyService: FamilyService, private incomeService: IncomeService) {
 
-    this.date = ""
+    this.date = new Date().toDateString();
     this.family = ""
     this.amount = ""
 
@@ -50,10 +50,12 @@ export class IncomeComponent implements OnInit {
     let familyInfo:any = []
     this.familyService.getFamily().subscribe((res: any) => {
      let arr = res
-     arr.forEach((element: any) => {
+     arr.forEach((family: any) => {
       // console.log(element)
+      if(family.status == "active"){
+        familyInfo.push(family)
 
-       familyInfo.push(element)
+      }
      });
 
     });
@@ -99,14 +101,19 @@ export class IncomeComponent implements OnInit {
   }
 
   deleteEntry(entry:any):void {
-    this.incomeService.deleteIncome(entry._id).subscribe((res: any) => {
-      this.getIncomes() // Once the record gets deleted we refetch
-    })
+    if(confirm("Are you sure you would like to delete this entry?")){
+      this.incomeService.deleteIncome(entry._id).subscribe((res: any) => {
+        this.getIncomes() // Once the record gets deleted we refetch
+      })
+    }
+    
   }
   updateEntry(entry:any):void {
-    this.incomeService.updateIncome(entry._id, entry).subscribe((res:any) => {
-      this.getIncomes() // After the record gets edited we refetch
-    })
+    if(confirm("Are you sure you would like to update this entry?")){
+      this.incomeService.updateIncome(entry._id, entry).subscribe((res:any) => {
+        this.getIncomes() // After the record gets edited we refetch
+     })
+   }
   }
   toggleEditEntry(entry:any){
     entry.isEditing = !entry.isEditing;
